requirements.txt
pandas
numpy
seaborn
scikit-learn
matplotlib
catboost
xgboost
dill
fastapi
uvicorn
python-multipart
pyyaml
# -e .